requirements.txt
streamlit
opencv-python-headless
ultralytics
easyocr
numpy
torch
Pillow